refactor(testing): pass cleanup func for truncating tables as argument to 'behaviour test functions'
After each test for repository, db should be cleaned up for the next tests to continue. You can now
create that cleanup func in repo_test.go and pass as argument to be used in interface behaviour
tests. Otherwise, you had to add cleanup function to the domain repository interface